feat : 스터디룸 생성 세부 작성

See merge request s11-webmobile1-sub2/S11P12B207!42

+++ backend/src/main/java/com/ssafy/ptpt/api/studyroom/response/StudyRoomListResponse.java
@@ -17,7 +17,7 @@ public class StudyRoomListResponse {
 
     private String studyRoomTitle;
 
-    private List<Integer> entryList;
+    private List<Long> entryList;
 
     private String studyRoomCode;
